src/recommendationservice/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
google-api-core==2.18.0
    # via -r requirements.in
google-auth==2.29.0
    # via google-api-core
googleapis-common-protos==1.63.0
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.62.1
    # via
    #   -r requirements.in
    #   grpcio-health-checking
    #   grpcio-tools
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio-health-checking==1.62.1
    # via -r requirements.in
grpcio-tools==1.62.1
    # via -r requirements.in
idna==3.7
    # via requests
importlib-metadata==7.0.0
    # via opentelemetry-api
opentelemetry-api==1.24.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
opentelemetry-exporter-otlp==1.24.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-common==1.24.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.24.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.24.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.45b0
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-grpc==0.45b0
    # via -r requirements.in
opentelemetry-proto==1.24.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.24.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation-grpc
opentelemetry-semantic-conventions==0.45b0
    # via
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-health-checking
    #   grpcio-tools
    #   opentelemetry-proto
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
python-json-logger==2.0.7
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-http
rsa==4.9
    # via google-auth
typing-extensions==4.11.0
    # via opentelemetry-sdk
urllib3==2.2.1
    # via requests
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
zipp==3.18.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
